requirements.txt
ssh2-python==1.0.0

